Se movió la función eventSignOut al archivo eventSignOut, se cambio el nombre de la función testCreate a createUser, se modificó el código de la función router, la función removeHash se movió a al archivo routes.js, se eliminaron las archivos getHash, clearHash, la carpeta templates, el archivo displayUserData.js, se le dió comportamiento al botón logout(se agregó botón)
