+++ src/lib/page/wall.js
@@ -8,7 +8,9 @@ const wall = () => {
       <img src="img/logoapp.png" alt="LogoCICLAPP" class="logoHorizontalWall">
       <p class="textLogoWall">CICLAPP</p>
     </div>
-      <i class="ri-logout-box-r-fill"></i>
+    <button class="exitButtonClass" id="exitButtonId">
+    <i class="ri-logout-box-r-fill"></i>
+    </button>
     </figure>
     <aside class="introUser"> 
     <h1 class="titleWall" id="titleId"> ¡Hola, Usuario!</h1>
@@ -33,7 +35,7 @@ const wall = () => {
 return wallOffPublication;
 };
 
-export const buttonP = () => {
+const buttonP = () => {
   const postForm = document.getElementById('postForm')
   const buttonP = document.getElementById('PostContentButton')
   if(buttonP){
@@ -43,4 +45,4 @@ export const buttonP = () => {
     postForm.reset();
   })
 }}
-export default wall;
+export { wall, buttonP };
