--- src/lib/templates/header.js
@@ -1,8 +1,0 @@
-
-const header = () => {
-  const viewHeader = ``;
-
-  return viewHeader;
-};
-
-export default header;
