+++ src/lib/utils/eventButtonGoogle.js
@@ -15,4 +15,4 @@ const eventButtonGoogle = () => {
     
 }}
 
-export default eventButtonGoogle;
+export { eventButtonGoogle };
