--- src/lib/utils/getHash.js
@@ -1,5 +1,0 @@
-const getHash = () => {
-  const hash = window.location.hash;
-  return hash;
-};
-export default getHash;
